requirements.txt
numpy==2.0.0
pandas==2.2.2
pillow==10.4.0
pyyaml==6.0.1
streamlit==1.36.0
torch==2.3.1
torchvision==0.18.1

